archive/dash_app_v1/requirements.txt
absl-py==0.10.0
actionlib==1.12.1
angles==1.9.12
apt-clone==0.2.1
apturl==0.5.2
arbotix_python==0.10.0
arbotix_sensors==0.10.0
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base_local_planner==1.16.7
beautifulsoup4==4.6.0
bleach==3.2.1
blinker==1.4
blob==0.16
bondpy==1.8.5
boto3==1.15.18
botocore==1.18.18
Brlapi==0.6.6
cachetools==4.1.1
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg-modules==0.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextvars==2.4
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cryptography==2.1.4
cupshelpers==1.0
cv_bridge==1.13.0
cycler==0.10.0
Cython==0.29.21
dash==2.9.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.8
ddynamic_reconfigure_python==0.0.1
decorator==4.1.2
defer==1.0.6
defusedxml==0.6.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro==1.7.0
distro-info===0.18ubuntu0.18.04.1
dlib==19.17.0
docutils==0.14
dynamic_reconfigure==1.6.5
easydict==1.10
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.2
feedparser==5.2.1
Flask==2.0.3
future==0.18.2
futures==3.1.1
gast==0.3.3
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gps_common==0.3.1
graphsurgeon==0.4.5
grpcio==1.33.0rc2
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
image_geometry==1.13.0
immutables==0.19
importlib-metadata==4.8.3
interactive-markers==1.11.5
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.13
itsdangerous==2.0.1
jedi==0.17.2
jetcam==0.0.0
jetson-stats==4.2.1
Jetson.GPIO==2.0.12
Jinja2==3.0.3
jmespath==0.10.0
joblib==0.11
joint_state_publisher==1.12.15
joint_state_publisher_gui==1.12.15
joy_listener==0.2.6
joy_mouse==0.1.15
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kdl-parser-py==1.13.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.2
MarkupSafe==2.0.1
matplotlib==3.3.2
message_filters==1.14.13
mistune==0.8.4
mock==3.0.5
moveit-core==1.0.10
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
nose==1.3.7
notebook==6.1.4
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.4
paho-mqtt==1.6.1
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pip-autoremove==0.9.1
plotly==5.14.1
pluggy==0.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycairo==1.16.2
pycocotools==2.0.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.7.1
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
pyserial==3.4
pytest==3.3.2
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python_qt_binding==0.4.4
pytz==2018.3
pyxdg==0.25
PyYAML==5.4.1
pyzmq==19.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resource_retriever==1.12.7
roman==2.0.0
rosapi==0.11.14
rosbag==1.14.13
rosboost-cfg==1.14.9
rosbridge-library==0.11.14
rosbridge-server==0.11.14
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
Rosmaster-Lib==1.7.3
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosserial_arduino==0.8.0
rosserial_client==0.8.0
rosserial_python==0.8.0
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rsa==4.6
rtabmap-ros==0.20.18
rviz==1.13.29
s3transfer==0.3.3
scikit-build==0.15.0
scikit-learn==0.19.1
scipy==1.4.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.12.8
Shapely==1.7.1
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smach_ros==2.0.1
smbus==1.1.post2
smbus2==0.4.2
smclib==1.8.5
srdfdom==0.5.2
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tenacity==8.2.2
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0+nv20.9
tensorflow-estimator==2.3.0
tensorrt==7.1.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf2_sensor_msgs==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.13
torch @ file:///home/nano/Documents/yolov7/torch-1.8.0-cp36-cp36m-linux_aarch64.whl
torch2trt==0.1.0
torchvision==0.9.0
tornado==6.0.4
tqdm==4.41.0
trac_ik_python==1.5.1
traitlets==4.3.3
trt-pose==0.0.1
turtlebot3_automatic_parking==1.1.0
turtlebot3_automatic_parking_vision==1.1.0
turtlebot3_autorace_camera==1.2.0
turtlebot3_autorace_control==1.2.0
turtlebot3_autorace_core==1.2.0
turtlebot3_autorace_detect==1.2.0
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing_extensions==4.1.1
ubuntu-drivers-common==0.0.0
uff==0.6.9
unique_id==1.0.6
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urdfdom-py==0.4.6
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wiimote==1.14.0
wrapt==1.12.1
xacro==1.13.17
xkit==0.0.0
yacs==0.1.8
zipp==3.3.1
zope.interface==4.3.2
